+++ packages/docs/package.json
@@ -12,18 +12,16 @@
 		"format": "prettier --plugin-search-dir . --write ."
 	},
 	"devDependencies": {
-		"@shikijs/rehype": "^1.5.1",
+		"@shikijs/rehype": "^1.5.2",
 		"@sveltejs/adapter-static": "^3.0.1",
 		"@sveltejs/kit": "^2.5.8",
 		"@sveltejs/vite-plugin-svelte": "^3.1.0",
-		"@types/d3-interpolate": "^3.0.4",
-		"@types/reveal.js": "^5.0.3",
-		"@typescript-eslint/eslint-plugin": "^7.8.0",
-		"@typescript-eslint/parser": "^7.8.0",
+		"@typescript-eslint/eslint-plugin": "^7.9.0",
+		"@typescript-eslint/parser": "^7.9.0",
 		"eslint": "^9.2.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.39.0",
-		"postcss-preset-env": "^9.5.12",
+		"postcss-preset-env": "^9.5.13",
 		"prettier": "^3.2.5",
 		"prettier-plugin-svelte": "^3.2.3",
 		"prettier-plugin-tailwindcss": "^0.5.14",
@@ -41,13 +39,13 @@
 		"vite": "^5.2.11"
 	},
 	"dependencies": {
-		"@animotion/motion": "^0.0.4",
+		"@animotion/core": "^0.0.5",
+		"@animotion/motion": "^0.0.5",
 		"@fontsource-variable/jetbrains-mono": "^5.0.21",
 		"@fontsource-variable/manrope": "^5.0.20",
 		"@tailwindcss/vite": "4.0.0-alpha.15",
-		"d3-interpolate": "^3.0.1",
 		"lucide-svelte": "^0.378.0",
 		"reveal.js": "^5.1.0",
-		"shiki": "^1.5.1"
+		"shiki": "^1.5.2"
 	}
 }
