+++ packages/docs/src/routes/docs/+page.md
@@ -4,4 +4,4 @@
 
 <iframe width="560" height="315" src="https://www.youtube.com/embed/dBfZjv-lt2k" title="Introducing Animotion" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" allowfullscreen></iframe>
 
-**Animotion** is a presentational framework for creating beautiful slides and visualizing ideas with code using [Svelte](https://svelte.dev/), [Reveal.js](https://revealjs.com/) and [Tailwind CSS](https://tailwindcss.com/).
+Animotion is a presentational framework for creating beautiful slides and visualizing ideas with code using [Svelte](https://svelte.dev/), [Reveal.js](https://revealjs.com/) and [Tailwind CSS](https://tailwindcss.com/).
