--- packages/docs/src/routes/docs/auto-animate/+page.server.ts
@@ -1,5 +1,0 @@
-import examples from './examples'
-
-export async function load() {
-	return { title: 'Auto-Animate', examples }
-}
