+++ packages/docs/src/routes/docs/auto-animate/+page.ts
@@ -1,0 +1,3 @@
+export async function load() {
+	return { title: 'Auto-Animate' }
+}
