+++ packages/docs/src/routes/docs/auto-animate/auto-animate.svelte
@@ -1,0 +1,18 @@
+<script lang="ts">
+	import { Embed as Presentation, Slide } from '@animotion/core'
+</script>
+
+<Presentation>
+	<Slide animate>
+		<p class="text-[100px]">Magic</p>
+	</Slide>
+
+	<Slide animate>
+		<p class="text-[200px] text-teal-300">Magic</p>
+	</Slide>
+
+	<Slide animate>
+		<p class="text-[200px]">🪄</p>
+		<p class="mt-[48px] text-[200px] text-teal-300">Magic</p>
+	</Slide>
+</Presentation>
