--- packages/docs/src/routes/docs/code-blocks/+page.server.ts
@@ -1,5 +1,0 @@
-import examples from './examples'
-
-export async function load() {
-	return { title: 'Code Blocks', examples }
-}
