+++ packages/docs/src/routes/docs/code-blocks/+page.ts
@@ -1,0 +1,3 @@
+export async function load() {
+	return { title: 'Code Blocks' }
+}
