+++ packages/docs/src/routes/docs/styles/+page.ts
@@ -1,0 +1,3 @@
+export async function load() {
+	return { title: 'Styles' }
+}
