+++ packages/docs/src/routes/docs/styles/tailwind.svelte
@@ -1,0 +1,27 @@
+<script lang="ts">
+	import { Embed as Presentation, Slide } from '@animotion/core'
+</script>
+
+<Presentation>
+	<Slide>
+		<div
+			class="mx-auto flex w-[800px] justify-between font-semibold text-gray-900"
+		>
+			<div
+				class="grid h-[240px] w-[240px] place-content-center rounded-full bg-red-400"
+			>
+				red
+			</div>
+			<div
+				class="grid h-[240px] w-[240px] place-content-center rounded-full bg-green-400"
+			>
+				green
+			</div>
+			<div
+				class="grid h-[240px] w-[240px] place-content-center rounded-full bg-blue-400"
+			>
+				blue
+			</div>
+		</div>
+	</Slide>
+</Presentation>
